+++ bin/gl-style-validate.ts
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-/* eslint-disable no-process-exit */
 
 import minimist from 'minimist';
 import rw from 'rw';