+++ src/expression/evaluation_context.ts
@@ -2,8 +2,18 @@ import {Color} from './values';
 import type {FormattedSection} from './types/formatted';
 import type {GlobalProperties, Feature, FeatureState} from './index';
 import {ICanonicalTileID} from '../tiles_and_coordinates';
+import {hasMultipleOuterRings} from '../util/classify_rings';
 
 const geometryTypes = ['Unknown', 'Point', 'LineString', 'Polygon'];
+const simpleGeometryType = {
+    'Unknown': 'Unknown',
+    'Point': 'Point',
+    'MultiPoint': 'Point',
+    'LineString': 'LineString',
+    'MultiLineString': 'LineString',
+    'Polygon': 'Polygon',
+    'MultiPolygon': 'Polygon'
+};
 
 class EvaluationContext {
     globals: GlobalProperties;
@@ -14,6 +24,7 @@ class EvaluationContext {
     canonical: ICanonicalTileID;
 
     _parseColorCache: {[_: string]: Color};
+    _geometryType: string;
 
     constructor() {
         this.globals = null;
@@ -29,8 +40,33 @@ class EvaluationContext {
         return this.feature && 'id' in this.feature ? this.feature.id : null;
     }
 
+    geometryDollarType() {
+        return this.feature ?
+            typeof this.feature.type === 'number' ? geometryTypes[this.feature.type] : simpleGeometryType[this.feature.type] :
+            null;
+    }
+
     geometryType() {
-        return this.feature ? typeof this.feature.type === 'number' ? geometryTypes[this.feature.type] : this.feature.type : null;
+        let geometryType = this.feature.type;
+        if (typeof geometryType !== 'number') {
+            return geometryType;
+        }
+        geometryType = geometryTypes[this.feature.type];
+        if (geometryType === 'Unknown') {
+            return geometryType;
+        }
+        const geom = this.geometry();
+        const len = geom.length;
+        if (len === 1) {
+            return geometryType;
+        }
+        if (geometryType !== 'Polygon') {
+            return `Multi${geometryType}`;
+        }
+        if (hasMultipleOuterRings(geom)) {
+            return 'MultiPolygon';
+        }
+        return 'Polygon';
     }
 
     geometry() {