+++ src/feature_filter/feature_filter.test.ts
@@ -249,7 +249,7 @@ describe('convert legacy filters to expressions', () => {
             [
                 'match',
                 ['geometry-type'],
-                ['LineString', 'Point', 'Polygon'],
+                ['LineString', 'MultiLineString', 'MultiPoint', 'MultiPolygon', 'Point', 'Polygon'],
                 true,
                 false
             ],
@@ -328,10 +328,96 @@ function legacyFilterTests(createFilterExpr) {
         expect(f({zoom: 0}, {properties: {}})).toBe(false);
     });
 
-    test('==, $type', () => {
-        const f = createFilterExpr(['==', '$type', 'LineString']).filter;
-        expect(f({zoom: 0}, {type: 1})).toBe(false);
-        expect(f({zoom: 0}, {type: 2})).toBe(true);
+    const UnknownGeometry = {type: 0};
+    const PointGeometry = {type: 1, geometry: [
+        [{x: 0, y: 0}]]};
+    const MultiPointGeometry = {type: 1, geometry: [
+        [{x: 0, y: 0}],
+        [{x: 1, y: 1}]]};
+    const LinestringGeometry = {type: 2, geometry: [
+        [{x: 0, y: 0}, {x: 1, y: 1}]]};
+    const MultiLineStringGeometry = {type: 2, geometry: [
+        [{x: 0, y: 0}, {x: 1, y: 1}],
+        [{x: 2, y: 0}, {x: 1, y: 0}]]};
+    const PolygonGeometry = {type: 3, geometry: [
+        [{x: 0, y: 0}, {x: 3, y: 0}, {x: 3, y: 3}, {x: 0, y: 3}, {x: 0, y: 0}],
+        [{x: 0, y: 0}, {x: 0, y: 2}, {x: 2, y: 2}, {x: 2, y: 0}, {x: 0, y: 0}]]};
+    const MultiPolygonGeometry = {type: 3, geometry: [
+        [{x: 0, y: 0}, {x: 3, y: 0}, {x: 3, y: 3}, {x: 0, y: 3}, {x: 0, y: 0}],
+        [{x: 0, y: 0}, {x: 0, y: 2}, {x: 2, y: 2}, {x: 2, y: 0}, {x: 0, y: 0}],
+        [{x: 0, y: 0}, {x: 1, y: 0}, {x: 1, y: 1}, {x: 0, y: 1}, {x: 0, y: 0}]]};
+
+    test('==, $type, Point', () => {
+        const fPoint = createFilterExpr(['==', '$type', 'Point']).filter;
+        expect(fPoint({zoom: 0}, UnknownGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, PointGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, MultiPointGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, LinestringGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, MultiLineStringGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, PolygonGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, MultiPolygonGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
+    });
+
+    test('==, $type, LineString', () => {
+        const fLineString = createFilterExpr(['==', '$type', 'LineString']).filter;
+        expect(fLineString({zoom: 0}, UnknownGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, PointGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, MultiPointGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, LinestringGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, MultiLineStringGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, PolygonGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, MultiPolygonGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
+    });
+
+    test('==, $type, Polygon', () => {
+
+        const fPolygon = createFilterExpr(['==', '$type', 'Polygon']).filter;
+        expect(fPolygon({zoom: 0}, UnknownGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, PointGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, MultiPointGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, LinestringGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, MultiLineStringGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, PolygonGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, MultiPolygonGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
+    });
+
+    test('==, $type, Unknown', () => {
+        const fUnknown = createFilterExpr(['==', '$type', 'Unknown']).filter;
+        expect(fUnknown({zoom: 0}, UnknownGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, PointGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, MultiPointGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, LinestringGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, MultiLineStringGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, PolygonGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, MultiPolygonGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
     });
 
     test('==, $id', () => {
@@ -373,10 +459,76 @@ function legacyFilterTests(createFilterExpr) {
         expect(f({zoom: 0}, {properties: {}})).toBe(true);
     });
 
-    test('!=, $type', () => {
-        const f = createFilterExpr(['!=', '$type', 'LineString']).filter;
-        expect(f({zoom: 0}, {type: 1})).toBe(true);
-        expect(f({zoom: 0}, {type: 2})).toBe(false);
+    test('!=, $type, Point', () => {
+        const fPoint = createFilterExpr(['!=', '$type', 'Point']).filter;
+        expect(fPoint({zoom: 0}, UnknownGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, PointGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, MultiPointGeometry)).toBe(false);
+        expect(fPoint({zoom: 0}, LinestringGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, MultiLineStringGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, PolygonGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, MultiPolygonGeometry)).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(fPoint({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(fPoint({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
+    });
+
+    test('!=, $type, LineString', () => {
+        const fLineString = createFilterExpr(['!=', '$type', 'LineString']).filter;
+        expect(fLineString({zoom: 0}, UnknownGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, PointGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, MultiPointGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, LinestringGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, MultiLineStringGeometry)).toBe(false);
+        expect(fLineString({zoom: 0}, PolygonGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, MultiPolygonGeometry)).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(fLineString({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(fLineString({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
+    });
+
+    test('!=, $type, Polygon', () => {
+        const fPolygon = createFilterExpr(['!=', '$type', 'Polygon']).filter;
+        expect(fPolygon({zoom: 0}, UnknownGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, PointGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, MultiPointGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, LinestringGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, MultiLineStringGeometry)).toBe(true);
+        expect(fPolygon({zoom: 0}, PolygonGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, MultiPolygonGeometry)).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(fPolygon({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(fPolygon({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
+    });
+
+    test('!=, $type, Unknown', () => {
+        const fUnknown = createFilterExpr(['!=', '$type', 'Unknown']).filter;
+        expect(fUnknown({zoom: 0}, UnknownGeometry)).toBe(false);
+        expect(fUnknown({zoom: 0}, PointGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, MultiPointGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, LinestringGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, MultiLineStringGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, PolygonGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, MultiPolygonGeometry)).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(fUnknown({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(fUnknown({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
     });
 
     test('<, number', () => {
@@ -563,15 +715,22 @@ function legacyFilterTests(createFilterExpr) {
 
     test('in, $type', () => {
         const f = createFilterExpr(['in', '$type', 'LineString', 'Polygon']).filter;
-        expect(f({zoom: 0}, {type: 1})).toBe(false);
-        expect(f({zoom: 0}, {type: 2})).toBe(true);
-        expect(f({zoom: 0}, {type: 3})).toBe(true);
+        expect(f({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
 
         const f1 = createFilterExpr(['in', '$type', 'Polygon', 'LineString', 'Point']).filter;
-        expect(f1({zoom: 0}, {type: 1})).toBe(true);
-        expect(f1({zoom: 0}, {type: 2})).toBe(true);
-        expect(f1({zoom: 0}, {type: 3})).toBe(true);
-
+        expect(f1({zoom: 0}, {type: 'Unknown'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'LineString'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'MultiLineString'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'Polygon'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'MultiPolygon'})).toBe(true);
     });
 
     test('!in, degenerate', () => {
@@ -621,9 +780,22 @@ function legacyFilterTests(createFilterExpr) {
 
     test('!in, $type', () => {
         const f = createFilterExpr(['!in', '$type', 'LineString', 'Polygon']).filter;
-        expect(f({zoom: 0}, {type: 1})).toBe(true);
-        expect(f({zoom: 0}, {type: 2})).toBe(false);
-        expect(f({zoom: 0}, {type: 3})).toBe(false);
+        expect(f({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'Point'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'MultiPoint'})).toBe(true);
+        expect(f({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(f({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
+
+        const f1 = createFilterExpr(['!in', '$type', 'Polygon', 'LineString', 'Point']).filter;
+        expect(f1({zoom: 0}, {type: 'Unknown'})).toBe(true);
+        expect(f1({zoom: 0}, {type: 'Point'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'MultiPoint'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'LineString'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'MultiLineString'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'Polygon'})).toBe(false);
+        expect(f1({zoom: 0}, {type: 'MultiPolygon'})).toBe(false);
     });
 
     test('any', () => {