+++ CHANGELOG.md
@@ -1,12 +1,33 @@
 ## main
 
 ### ✨ Features and improvements
-
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
-
 - _...Add new stuff here..._
+
+## 21.2.0
+
+### ✨ Features and improvements
+Add `vertical-perspective` projection ([#890](https://github.com/maplibre/maplibre-style-spec/pull/890))
+
+## 21.1.0
+
+### ✨ Features and improvements
+
+- Added `centerAltitude` property to stylesheet ([#851](https://github.com/maplibre/maplibre-style-spec/issues/851))
+
+## 21.0.0
+
+### ✨ Features and improvements
+
+- Aligned the implementation of `["geometry-type"]` with [its spec](https://maplibre.org/maplibre-style-spec/expressions/#geometry-type). Now, when applicable, `["geometry-type"]` returns values not available while using `"$type"`: `"MultiPoint"`, `"MultiLineString"`, and `"MultiPolygon"`. The behaviour of `"$type"` has not changed. ([#519](https://github.com/maplibre/maplibre-style-spec/pull/519))
+
+## 20.4.0
+
+### ✨ Features and improvements
+
+* Added `roll` property to stylesheet ([#850](https://github.com/maplibre/maplibre-style-spec/issues/850))
 
 ## 20.3.1